refactor(hooks): split hooks.go to satisfy file length lint
